Reviewed EventsToFrame ProcessModel and unit tests.
